Fix NS04 replay gate
